refactor(pdf): isolate bounded algorithms
